Changed the format of submitted grades from Int to Float

+++ app/src/main/java/com/sweak/diplomaexam/data/remote/dto/session/SubmitAdditionalGradesRequestDto.kt
@@ -2,7 +2,7 @@ package com.sweak.diplomaexam.data.remote.dto.session
 
 data class SubmitAdditionalGradesRequestDto(
     val sessionId: Int,
-    val diplomaGrade: Int,
-    val presentationGrade: Int,
-    val studyGrade: Int
+    val diplomaGrade: Float,
+    val presentationGrade: Float,
+    val studyGrade: Float
 )
